Implement the Gitter support
Closes #117 (Gitter.im chat support)

So the downside of this solution is that you cannot have nice red
background anymore. That is, you cannot control so-called "error level".
If Bugsnag was implemented as a Gitter service[1], then it would have
its own icon, too.

According to the docs[2] we can pass "html". So in theory it's probably
possible to set background in that way, but it didn't work for me and
I'm not sure why.

[1] https://github.com/gitterHQ/services/pull/26
[2] https://developer.gitter.im/docs/messages-resource